% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canary-package.R
\docType{package}
\name{canary-package}
\alias{canary}
\alias{canary-package}
\title{canary: Check That Build Tools are Working}
\description{
Checks that build tools such as Travis, Appveyor and R-hub are
working, before you look for bugs in your own package
}
\author{
\strong{Maintainer}: Duncan Garmonsway \email{duncan.garmonsway@gmail.com}

}
\keyword{internal}
